(PC-17811) CulturalSurvey: put shared helpers in /shared folders
