+++ src/features/culturalSurvey/helpers/utils.ts
@@ -1,9 +1,4 @@
-import {
-  CulturalSurveyQuestionEnum,
-  CulturalSurveyQuestionsResponse,
-  UserProfileResponse,
-} from 'api/gen/api'
-import { useSettingsContext } from 'features/auth/context/SettingsContext'
+import { CulturalSurveyQuestionEnum, CulturalSurveyQuestionsResponse } from 'api/gen/api'
 
 export const mapQuestionIdToPageTitle = (id: CulturalSurveyQuestionEnum | undefined) => {
   switch (id) {
@@ -16,23 +11,6 @@ export const mapQuestionIdToPageTitle = (id: CulturalSurveyQuestionEnum | undefi
     default:
       return ''
   }
-}
-
-export function useCulturalSurveyRoute() {
-  const { data: settings } = useSettingsContext()
-
-  // as long as ENABLE_CULTURAL_SURVEY is still used in the api to define if user should
-  // fill cultural survey, we use enableNativeCulturalSurvey to know which cultural survey
-  // (native or typeform) to show
-  if (settings?.enableNativeCulturalSurvey) {
-    return 'CulturalSurveyIntro'
-  }
-  return 'CulturalSurvey'
-}
-
-export function shouldShowCulturalSurvey(user?: UserProfileResponse) {
-  if (!user) return false
-  return user.needsToFillCulturalSurvey
 }
 
 export const createInitialQuestionsList = (
