+++ src/features/firstLogin/helpers.tsx
@@ -6,13 +6,13 @@ import { v1 as uuidv1 } from 'uuid'
 
 import { api } from 'api/api'
 import { useAuthContext } from 'features/auth/context/AuthContext'
-import { shouldShowCulturalSurvey } from 'features/culturalSurvey/helpers/utils'
 import { navigateToHome, useCurrentRoute } from 'features/navigation/helpers'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { homeNavConfig } from 'features/navigation/TabBar/helpers'
 import { env } from 'libs/environment'
 import { MonitoringError } from 'libs/monitoring'
 import { QueryKeys } from 'libs/queryKeys'
+import { shouldShowCulturalSurvey } from 'shared/culturalSurvey/shouldShowCulturalSurvey'
 import { LoadingPage } from 'ui/components/LoadingPage'
 
 const FORM_ID = env.CULTURAL_SURVEY_TYPEFORM_ID
