+++ src/shared/culturalSurvey/shouldShowCulturalSurvey.ts
@@ -1,0 +1,6 @@
+import { UserProfileResponse } from 'api/gen'
+
+export function shouldShowCulturalSurvey(user?: UserProfileResponse) {
+  if (!user) return false
+  return user.needsToFillCulturalSurvey
+}
